Move API serializers to v1
