+++ lunes_cms/api/v1/serializers/__init__.py
@@ -1,0 +1,12 @@
+"""
+This module contains the model serializers, see :doc:`django:topics/serialization`.
+"""
+
+from .alternative_word_serializer import AlternativeWordSerializer
+from .discipline_serializer import DisciplineSerializer
+from .document_image_list_serializer import DocumentImageListSerializer
+from .document_image_serializer import DocumentImageSerializer
+from .document_serializer import DocumentSerializer
+from .feedback_serializer import FeedbackSerializer
+from .group_serializer import GroupSerializer
+from .training_set_serializer import TrainingSetSerializer
