+++ lunes_cms/api/v1/serializers/alternative_word_serializer.py
@@ -1,0 +1,18 @@
+from rest_framework import serializers
+
+from ....cms.models import AlternativeWord
+
+
+class AlternativeWordSerializer(serializers.ModelSerializer):
+    """
+    Serializer for the AlternativeWord module. Inherits from
+    `serializers.ModelSerializer`.
+    """
+
+    class Meta:
+        """
+        Define model and the corresponding fields
+        """
+
+        model = AlternativeWord
+        fields = ("alt_word", "article")
