+++ lunes_cms/api/v1/views/discipline_filtered_viewset.py
@@ -1,13 +1,13 @@
 from rest_framework import mixins, viewsets
 
 from ....cms.models import Discipline
-from ...serializers import DisciplineSerializer
 from ...utils import (
     get_filtered_discipline_queryset,
     get_discipline_by_group_queryset,
     get_overview_discipline_queryset,
     check_group_object_permissions,
 )
+from ..serializers import DisciplineSerializer
 
 
 class DisciplineFilteredViewSet(mixins.ListModelMixin, viewsets.GenericViewSet):
